1: Move test to conv_sum_broadcast.cpp; 2: Add a stricter condition to change minDim to 1;
